requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-core
langchain-huggingface
pinecone
pinecone-text
sentence-transformers
scikit-learn
pinecone-notebooks